shifted validation code into shared folder
